Remov console error + minor refactor
https://coveord.atlassian.net/browse/JSUI-2471
